requirements.txt
coverage==5.5
cycler==0.10.0
dj-database-url==0.5.0
Django==2.2.3
gunicorn==20.1.0
kiwisolver==1.3.1
matplotlib==3.4.1
mysqlclient==2.0.3
numpy==1.20.2
Pillow==8.2.0
psycopg2==2.8.6
pyparsing==2.4.7
python-dateutil==2.8.1
python-decouple==3.4
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
whitenoise==5.2.0
